r/aws_glue_registry: Reduce visibility.
